Remove unused params after refactoring `commit-generated-manifest` task
`commit-generated-manifest` task previously copied both manifest and
opsfile in the script no matter what. That made those params be required
in the task and forced us to pass dummy values in pipeline definitions.

Recently, `cp` has started to fail due to the dummy source file and dummy
destination file being the *same file*.  We do not know what caused the
change in behavior, but it has forced us to do the right thing, and only
conditionally carry out the `cp` if it's called for in the job.

[#146645305]

Signed-off-by: Chunyi Lyu <clyu@pivotal.io>