Update releases in compiled releases ops file one by one when stemcell is bumped
Previously, we had a task with hard-coded list of releases that were updated in the
corresponding use-compiled-releases*.yml ops-file. This is not great, because each time
we add a release to the base manifest and forget to update the list of releases in the task, the compiled releases
ops file will not be updated correctly. This will cause unit-tests to fail in the pipeline. [#159421027]

Signed-off-by: Iryna Shustava <ishustava@pivotal.io>