Don't ignore directory if files are ignored
If an ignored file is found when walking cf-deployment directory,
we used to return filepath.SkipDir, which will ignore the whole directory.
This is undesirable because sometimes this directory may contain other subdirectories
with ops files we'd like to find. We now instead separate ignoreFiles list from
ignoreDirs list and only return filepath.SkipDir for ignored directories

[#160007374]