update-manifest-releases should ignore .github dir
Add .github to the list of ignored directories in cf-d.

Recently added a .github folder to the cf-deployment repo, and it
contains yaml files that were pulled in by this util. This util would
then fail because the yaml files in that folder are not ops files, and
therefore are not in the format that this util expected them to be in.