CI: update destroy-infrastructure-lite job (#216)
In most cases, we need to delete deployments before destroying the
director so that we don't leave orphaned VMs lying around. However,
lite is a bosh-lite, which reuses the director VM for any bosh
deployments.

In this case, we should not delete deployments before destroying the
director then. Removing that step should speed this operation up. Also,
we won't have to manually remove it when trying to destroy the env
when the director is unresponsive.

Signed-off-by: Carson Long <lcarson@vmware.com>
Co-authored-by: Dave Walter <walterda@vmware.com>